Increase size of sprtopscreen
Fixes a potential overflow due to the "wiggle fix".

+++ src/r_segs.c
@@ -204,7 +204,7 @@ static void R_DrawMaskedColumn(column_t *column)
 {
     while (column->topdelta != 0xff)
     {
-        int     topscreen;
+        int64_t topscreen;
 
         if (column->length == 0)
         {
@@ -215,8 +215,8 @@ static void R_DrawMaskedColumn(column_t *column)
         // calculate unclipped screen coordinates for post
         topscreen = sprtopscreen + spryscale * column->topdelta + 1;
 
-        dc_yl = MAX((topscreen + FRACUNIT) >> FRACBITS, mceilingclip[dc_x] + 1);
-        dc_yh = MIN((topscreen + spryscale * column->length) >> FRACBITS, mfloorclip[dc_x] - 1);
+        dc_yl = MAX((int)((topscreen + FRACUNIT) >> FRACBITS), mceilingclip[dc_x] + 1);
+        dc_yh = MIN((int)((topscreen + spryscale * column->length) >> FRACBITS), mfloorclip[dc_x] - 1);
 
         dc_texturefrac = dc_texturemid - (column->topdelta << FRACBITS) +
             FixedMul((dc_yl - centery) << FRACBITS, dc_iscale);
@@ -309,7 +309,7 @@ void R_RenderMaskedSegRange(drawseg_t *ds, int x1, int x2)
                 dc_colormap = walllights[BETWEEN(0, spryscale >> LIGHTSCALESHIFT,
                     MAXLIGHTSCALE - 1)];
 
-            sprtopscreen = (long)(t >> FRACBITS);
+            sprtopscreen = (int64_t)(t >> FRACBITS);
             dc_iscale = 0xffffffffu / (unsigned int)spryscale;
 
             // draw the texture
